requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
anyio==4.9.0
attrs==25.3.0
babel==2.17.0
Brotli==1.1.0
certifi==2025.4.26
charset-normalizer==3.4.2
docxcompose==1.4.0
docxtpl==0.20.0
et_xmlfile==2.0.0
frozenlist==1.6.0
g4f==0.5.2.9
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.10
Jinja2==3.1.6
lxml==5.4.0
MarkupSafe==3.0.2
multidict==6.4.4
nest-asyncio==1.6.0
openpyxl==3.1.5
pillow==11.2.1
propcache==0.3.1
pycryptodome==3.23.0
PyPDF2==3.0.1
python-docx==1.1.2
python-telegram-bot==20.0
requests==2.32.3
rfc3986==1.5.0
setuptools==80.8.0
six==1.17.0
sniffio==1.3.1
typing_extensions==4.13.2
urllib3==2.4.0
yarl==1.20.0
